Function to help with synchronization
Added getInputFramesBuffered(), which can be used for phase synchronization using a PLL. We needed some way to relate the input to the output. Ideally the output buffer could tag each range with the stretch ratio that produced it. Instead, approximate using the current stretch ratio, and the results are good enough for general synchronizing. The amount of time required to synchronize is also dependent on the lowpass filter in the PLL, but good results are achieved usually within about 5 to 8 blocks of processing and the phase error diminishes to zero.

To get an approximation of the total frames in the stretcher, we need only to expose how many frames are in the stretchers input buffer, since available() will return the number of frames in the output buffer already. To get the total frames we then use available() divided by getTimeRatio() and then add getInputFramesBuffered().

Something like:

Float64 TimestretcherRB::_framesInStretcher() {

	// How many frames in the output buffer
	int64_t outBuffer = _rubber->available();

	// Ratio of approximately how these output frames were stretched
	Float64 outRatio = _rubber->getTimeRatio();

	// Convert to number of frames in input buffer 'space'
	Float64 outSpace = outBuffer / (outRatio ? : 1.0);

	// How many frames in the input buffer
	int64_t inSpace = _rubber->getInputFramesBuffered();

	// Total buffered in stretcher, as measured in INPUT space
	Float64 totalBuffered = inSpace + outSpace;

	return totalBuffered;
}